feat: blogContent component
